refactor(cli): extract subagent runtime
