+++ packages/agent-cli/agent-cli.cabal
@@ -63,6 +63,7 @@ library
                     , Agent.CLI.Progress
                     , Agent.CLI.Project
                     , Agent.CLI.Prompt
+                    , Agent.CLI.Request
                     , Agent.CLI.ProviderFallback
                     , Agent.CLI.ProviderTransition
                     , Agent.CLI.Render
@@ -73,6 +74,7 @@ library
                     , Agent.CLI.SessionTitle
                     , Agent.CLI.Status
                     , Agent.CLI.SubagentStore
+                    , Agent.CLI.Subagents.Runtime
                     , Agent.CLI.Style
                     , Agent.CLI.Terminal
                     , Agent.CLI.Timestamp
@@ -146,6 +148,7 @@ test-suite agent-cli-test
                     , Agent.CLI.PromptSpec
                     , Agent.CLI.ProviderFallbackSpec
                     , Agent.CLI.ProviderTransitionSpec
+                    , Agent.CLI.RequestSpec
                     , Agent.CLI.RenderSpec
                     , Agent.CLI.ReplStatusSpec
                     , Agent.CLI.ResumeSpec
