+++ packages/agent-cli/test/Spec.hs
@@ -31,6 +31,7 @@ import qualified Agent.CLI.ProjectSpec as ProjectSpec
 import qualified Agent.CLI.PromptSpec as PromptSpec
 import qualified Agent.CLI.ProviderFallbackSpec as ProviderFallbackSpec
 import qualified Agent.CLI.ProviderTransitionSpec as ProviderTransitionSpec
+import qualified Agent.CLI.RequestSpec as RequestSpec
 import qualified Agent.CLI.RenderSpec as RenderSpec
 import qualified Agent.CLI.ReplStatusSpec as ReplStatusSpec
 import qualified Agent.CLI.ResumeSpec as ResumeSpec
@@ -75,6 +76,7 @@ main = hspec do
     PromptSpec.spec
     ProviderFallbackSpec.spec
     ProviderTransitionSpec.spec
+    RequestSpec.spec
     RenderSpec.spec
     ReplStatusSpec.spec
     ResumeSpec.spec
